# openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Ants API
  version: 1.0.0
  description: |
    Public HTTP API for the Ants software-engineering agent platform,
    versioned under `/v1`.

    Contract rules carried over from the master plan:
    - Every remotely retryable mutation requires an `Idempotency-Key`.
    - Errors use RFC 9457 Problem Details (`application/problem+json`)
      with a stable `code` extension field.
    - Event streams are paginated by a monotonic `seq` cursor; clients
      pass `after=<last seen seq>` to resume without duplicates.
    - Tenant scoping is implicit and mandatory: every authenticated
      request operates inside exactly one tenant.

    Authentication: servers act as OIDC resource servers (ADR-0019).
    Authenticated routes require `Authorization: Bearer <access token>` — a
    JWT issued by the configured identity provider, verified for signature,
    issuer, audience, and validity window; the tenant and subject are derived
    exclusively from verified claims. Without a configured provider every
    authenticated route refuses with problem code
    `authentication_not_configured`. There is no development bypass.
servers:
  - url: http://127.0.0.1:8080
    description: Local development default
...
paths:
...
  /v1/threads/{id}/runs:
    get:
      tags: [runs]
      summary: List the runs of a thread
      description: |
        Runs are returned oldest first in one stable order (creation time,
        ties broken by ascending id). Runs are never deleted and new runs
        arrive only at the tail, so positions never reshuffle: pages fetched
        with increasing `after` cursors never duplicate or skip, even when
        concurrent starts grow `total` between page requests. The newest run
        is therefore the last item of the final page, not of the first one —
        clients wanting the true latest must walk pages until they have
        consumed `total` entries (the console does exactly this; it must not
        assume any single bounded page holds the latest run). A known thread
        without runs yields an empty `runs` array, never an error; an unknown
        or foreign-tenant thread yields the uniform 404 problem (no existence
        oracle, ADR-0004).
      operationId: listThreadRuns
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ThreadID"
        - $ref: "#/components/parameters/AfterCursor"
      responses:
        "200":
          description: One bounded page of the thread's runs with the total count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunPage"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    post:
      tags: [runs]
      summary: |
        Start a run for the thread. Requires an `Idempotency-Key`;
        retrying the same call replays the same run instead of creating one.
      operationId: startRun
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ThreadID"
        - name: Idempotency-Key
          in: header
          required: true
          maxLength: 256
          schema:
            type: string
      responses:
        "202":
          description: Run accepted; execution proceeds asynchronously
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token issued by the deployment's identity provider
        (Keycloak locally). Tokens must be signed with RS256, carry the API's
        audience, name the Ants tenant in the configured claim (default
        `ants_tenant`), and stay within their validity window. Rejections use
        stable problem codes: missing_bearer_token,
        invalid_authorization_header, unsupported_token_algorithm,
        malformed_token, invalid_token_signature, untrusted_issuer,
        audience_mismatch, token_expired, token_not_yet_valid,
        missing_subject, missing_exp, invalid_tenant_claim, unknown_tenant;
        identity-provider outages surface as transient problems instead.
  parameters:
    ThreadID:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ThreadID" }
...
    AfterCursor:
      name: after
      in: query
      required: false
      description: |
        Return entries at positions strictly greater than this value in the
        stable oldest-first order. The grammar is exactly this schema
        (integer, minimum 0, default 0): omitted means 0; a present value
        must be a single decimal-digit integer (no sign, whitespace, or
        other padding) within the int64 range. Repeated, empty, or malformed
        values are rejected with problem code `invalid_cursor` rather than
        silently falling back to the default.
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
  responses:
    Problem:
      description: Error as RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      required: [type, code, title, status]
      properties:
        type: { type: string, format: uri }
        code:
          type: string
          description: Stable machine-readable identifier for the failure mode
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        extras:
          type: object
          additionalProperties: true
    TenantID: { type: string, pattern: "^ten_[A-Za-z0-9]{20,64}$" }
...
    ThreadID: { type: string, pattern: "^thr_[A-Za-z0-9]{20,64}$" }
...
    RunID: { type: string, pattern: "^run_[A-Za-z0-9]{20,64}$" }
    TaskID: { type: string, pattern: "^tsk_[A-Za-z0-9]{20,64}$" }
    SpecID: { type: string, pattern: "^spc_[A-Za-z0-9]{20,64}$" }
    ArtifactID: { type: string, pattern: "^art_[A-Za-z0-9]{20,64}$" }
...
    Timestamp:
      type: string
      format: date-time
...
    FailureInfo:
      type: object
      required: [code, message, transient]
      properties:
        code: { type: string }
        message: { type: string }
        transient: { type: boolean }
    TaskStatus:
      type: string
      enum:
        - draft
        - queued
        - provisioning
        - working
        - verifying
        - integrating
        - done
        - waiting_external
        - blocked
        - cancelled
        - failed
    RunStatus:
      type: string
      enum:
        - pending
        - planning
        - executing_tasks
        - integrating
        - verifying
        - reporting
        - completed
        - failed
        - cancelled
...
    Run:
      type: object
      required: [id, tenant_id, thread_id, status, idempotency_key, task_ids, version, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/RunID" }
        tenant_id: { $ref: "#/components/schemas/TenantID" }
        thread_id: { $ref: "#/components/schemas/ThreadID" }
        spec_id: { $ref: "#/components/schemas/SpecID" }
        status: { $ref: "#/components/schemas/RunStatus" }
        idempotency_key: { type: string }
        task_ids:
          type: array
          items: { $ref: "#/components/schemas/TaskID" }
        report:
          $ref: "#/components/schemas/RunReport"
        failure: { $ref: "#/components/schemas/FailureInfo" }
        version: { type: integer, format: int64 }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
        finished_at: { $ref: "#/components/schemas/Timestamp" }
...
    RunPage:
      type: object
      description: One page of a thread's run history in stable creation order
      required: [runs, total]
      properties:
        runs:
          type: array
          items: { $ref: "#/components/schemas/Run" }
        total:
          type: integer
          format: int64
          minimum: 0
          description: Total number of runs recorded for the thread
...
    SpecContent:
      type: object
      required: [outcome, assumptions, requirements, non_goals, success_criteria, blockers]
      properties:
        outcome: { type: string }
        assumptions:
          type: array
          items: { type: string }
        requirements:
          type: array
          items: { type: string }
        non_goals:
          type: array
          items: { type: string }
        success_criteria:
          type: array
          items: { type: string }
        blockers:
          type: array
          items: { type: string }
    Severity: { type: string, enum: [blocker, warning, note] }
    Confidence: { type: string, enum: [high, medium] }
    Finding:
      type: object
      required: [category, severity, confidence, location, scenario]
      properties:
        category: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        location: { type: string }
        scenario:
          type: string
          description: Concrete failure scenario the finding describes
    Evidence:
      type: object
      required: [criterion, command, exit_code, passed, at]
      properties:
        criterion: { type: string }
        command:
          type: array
          items: { type: string }
        exit_code: { type: integer }
        passed: { type: boolean }
        log_artifact_id: { $ref: "#/components/schemas/ArtifactID" }
        at: { $ref: "#/components/schemas/Timestamp" }
    ReportTask:
      type: object
      required: [id, name, branch, status, attempts]
      properties:
        id: { $ref: "#/components/schemas/TaskID" }
        name: { type: string }
        branch: { type: string }
        commit_sha: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        attempts: { type: integer }
        failure: { $ref: "#/components/schemas/FailureInfo" }
    ReportIntegration:
      type: object
      required: [branch]
      properties:
        branch: { type: string }
        sha: { type: string }
        conflicts:
          type: array
          items: { type: string }
    ReportVerification:
      type: object
      required: [passed, evidence]
      properties:
        passed: { type: boolean }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/Evidence" }
    BudgetSummary:
      type: object
      required: [tasks_used, max_tasks, exec_ops_used, max_exec_ops]
      properties:
        tasks_used: { type: integer }
        max_tasks: { type: integer }
        exec_ops_used: { type: integer }
        max_exec_ops: { type: integer }
    ReportArtifactRef:
      type: object
      required: [id, kind, digest]
      properties:
        id: { $ref: "#/components/schemas/ArtifactID" }
        kind: { type: string, enum: [diff, log, report] }
        digest: { type: string }
    RunReport:
      type: object
      required: [run_id, thread_id, spec_id, spec, tasks, integration, verification, findings, artifacts, budget, ready_for_review, summary, started_at, finished_at]
      properties:
        run_id: { $ref: "#/components/schemas/RunID" }
        thread_id: { $ref: "#/components/schemas/ThreadID" }
        spec_id: { $ref: "#/components/schemas/SpecID" }
        spec: { $ref: "#/components/schemas/SpecContent" }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/ReportTask" }
        integration: { $ref: "#/components/schemas/ReportIntegration" }
        verification: { $ref: "#/components/schemas/ReportVerification" }
        findings:
          type: array
          items: { $ref: "#/components/schemas/Finding" }
        artifacts:
          type: array
          items: { $ref: "#/components/schemas/ReportArtifactRef" }
        budget: { $ref: "#/components/schemas/BudgetSummary" }
        ready_for_review: { type: boolean }
        summary: { type: string }
        started_at: { $ref: "#/components/schemas/Timestamp" }
        finished_at: { $ref: "#/components/schemas/Timestamp" }